pyproject.toml
[tool.poetry]
name = "dotpyle"
version = "0.1.0"
description = "Test"
authors = ["Jorge Bodega <jorge.bodega.f@gmail.com>", "Perseo Gutiérrez <>", ""]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
flake8 = "^3.8.4"
pylint = "^2.6.0"
click = "^7.1.2"
GitPython = "^3.1.12"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
pylint = "^2.6.0"
pep8 = "^1.7.1"
pre-commit = "^2.9.3"
pytest-cov = "^2.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
app = "dotpyle.main:main"
test = "pytest:main"

[tool.black]
line-length = 88
target-version = ['py36', 'py37', 'py38']
exclude = '''
/(
  | \.git
  | \.venv
  | buck-out
  | build
  | dist
)/
'''
verbose = true
